前回登板のリセットをbulk_updateで行うように修正 PGにイニングを表示し、テキストの色をつけるように修正

+++ eikan/models/pitcher_total_results.py
@@ -195,9 +195,9 @@ class PitcherTotalResults(models.Model):
         editable=False,
     )
 
-    previous_game_pitched = models.PositiveIntegerField(
+    previous_game_pitched = models.FloatField(
         verbose_name="前回登板",
-        default=0,
+        default=0.0,
         editable=False,
     )
